metrika -> metrica
